move language extension with others
