+++ src/Configuration/Models/ConfigurationDefinitions.cs
@@ -74,5 +74,20 @@
                 _ => throw new NotImplementedException()
             };
         }
+
+        public static string GetRiotRegionCode(this ProgramLanguage language)
+        {
+            return language switch
+            {
+                ProgramLanguage.En => "en_US",
+                ProgramLanguage.ZhHans => "zh_CN",
+                ProgramLanguage.ZhHant => "zh_TW",
+                ProgramLanguage.De => "de_DE",
+                ProgramLanguage.Es => "es_ES",
+                ProgramLanguage.Fr => "fr_FR",
+                ProgramLanguage.Pt => "pt_BR",
+                _ => "en_US",
+            };
+        }
     }
 }
