+++ src/UI.Main/Pages/WelcomeSetupDownload.xaml.cs
@@ -1,4 +1,5 @@
-﻿using Fraxiinus.ReplayBook.StaticData.Extensions;
+﻿using Fraxiinus.ReplayBook.Configuration.Models;
+using Fraxiinus.ReplayBook.StaticData.Extensions;
 using Fraxiinus.ReplayBook.StaticData.Models;
 using Fraxiinus.ReplayBook.UI.Main.Models;
 using Fraxiinus.ReplayBook.UI.Main.Utilities;
